core/tests/unit_tests/orchestrator/test_secure_aggregation.py
import unittest
from core.orchestrator.secure_aggregation import SecureAggregation
from unittest.mock import patch, MagicMock

class TestSecureAggregation(unittest.TestCase):
    def setUp(self):
        # Set up an instance of the SecureAggregation before each test
        self.secure_aggregator = SecureAggregation()

    @patch("core.orchestrator.secure_aggregation.SecureAggregation.aggregate_encrypted_updates")
    def test_aggregate_encrypted_updates(self, mock_aggregate_encrypted_updates):
        # Test aggregation of encrypted model updates
        mock_aggregate_encrypted_updates.return_value = "Aggregation successful"
        result = self.secure_aggregator.aggregate_encrypted_updates(["update_1", "update_2"])
        mock_aggregate_encrypted_updates.assert_called_once_with(["update_1", "update_2"])
        self.assertEqual(result, "Aggregation successful")

    @patch("core.orchestrator.secure_aggregation.SecureAggregation.apply_differential_privacy")
    def test_apply_differential_privacy(self, mock_apply_differential_privacy):
        # Test applying differential privacy to aggregated model
        mock_apply_differential_privacy.return_value = "Differential privacy applied"
        result = self.secure_aggregator.apply_differential_privacy("aggregated_model")
        mock_apply_differential_privacy.assert_called_once_with("aggregated_model")
        self.assertEqual(result, "Differential privacy applied")

    @patch("core.orchestrator.secure_aggregation.SecureAggregation.verify_secure_aggregation")
    def test_verify_secure_aggregation(self, mock_verify_secure_aggregation):
        # Test verification of the secure aggregation process
        mock_verify_secure_aggregation.return_value = True
        verification_result = self.secure_aggregator.verify_secure_aggregation("aggregated_model")
        mock_verify_secure_aggregation.assert_called_once_with("aggregated_model")
        self.assertTrue(verification_result)

    def tearDown(self):
        # Clean up after each test
        self.secure_aggregator = None

if __name__ == "__main__":
    unittest.main()
